ジェネリック実装作ってみたけど微妙

+++ Demo/2016/Unicode/UtfString/Utf8/ArrayAccessor.cs
@@ -6,7 +6,7 @@
 
         public ArrayAccessor(byte[] data) { _data = data; }
 
-        public ushort this[int index] => _data[index];
+        public byte this[int index] => _data[index];
 
         public int Length => _data.Length;
     }
